requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.8,<3.0
drf-spectacular>=0.27.0,<0.28
django-tailwind[reload]
python-dotenv>=0.19.0,<1.0

